fix makefile to use order-only prerequisites
this way the obj directory only gets checked for existence - we create it if it doesn't exist but not otherwise. not bothering to check timestamps is the key change, and this corrects a bug in the makefile that it would rebuild everything every time because of a loop of updating the timestamp on the obj directory requiring new objects be built requiring the obj directory to be refreshed... anyway, fixed now. (bug introduced when I made the makefile responsible for creating the obj directory)