broadwayd workaround for linux workflow
an instance of broadway is started in the background and used as the display server for the testbench process. this should get around the issue where gtk complains of not being able to connect to any window server, which causes the process to exit prematurely without actually running any tests.