add informative dropdown text
when a map is successfully loaded via the Open From File dropdown option or CLI args, we go back to insert a dropdown option with the file's name and set the value to that. also, add g_signal_handlers_block_by_func calls to prevent the duplicated calls with non-null invalid mapnames which this would otherwise generate which would break the system by trying to load another map on top of the one which has already been loaded (also removes some harmless but wasteful duplicated calls which used to occur).