fix a leak in the leak checker
to be fair, it was the only one that could possibly leak due to not
being tracked within the csa_alloc system, and still has to be allocated
for 100% of runtime, so is barely worth freeing. But this officially
means I can slap that "Leak-Free!" label back on it. If I ignore the
several MB of leaks due to GTK. Y'know. They're not my fault.